[C] Refactor calling convention
Big refactor.  Every argument to every `Function` (including methods) is
now of type `Object`, which comprises all values, modules, and
references.

Values, modules, and references are all of different kinds, but will
appear in similar places in source code.  Unifying them makes it
possible to assert the kinds of parameters, which is nice because the C
type system is useless with varargs.

The same unification would be necessary for a stack calling convention.

This change also gives `fn example(self)` and `fn example(^self)`
different type signatures, instead of always taking the receiver by
reference.  This will prevent a hella subtle bug far in the future of
this project:

    impl A {
        fn method(self, _: Unit) { ... }
    }

    a->method({ a := 1 })

    # If the method takes the receiver by value, it MUST be copied.
    # If it were not copied, the modification would make the new value
    #    the receiver.
    # No one would ever write this code but it's nice to be thorough.